fix: rename folder so test and lib match.
Fargate patterns were in a folder named ’fargate’, but EC2 patterns were in a folder named ’ecs’. In the test folder the naming was ’ec2’, so rename in lib to make it match. Note that this is an internal implementation detail only, it doesn’t affect public api.
