+++ packages/@aws-cdk/aws-ecs-patterns/lib/ec2/application-load-balanced-ecs-service.ts
