+++ packages/@aws-cdk/aws-ecs-patterns/lib/ec2/network-load-balanced-ecs-service.ts
