+++ packages/@aws-cdk/aws-ecs-patterns/lib/ec2/network-multiple-target-groups-ecs-service.ts
