+++ packages/@aws-cdk/aws-ecs-patterns/lib/ec2/queue-processing-ecs-service.ts
