+++ packages/@aws-cdk/aws-ecs-patterns/lib/ec2/scheduled-ecs-task.ts
