+++ packages/@aws-cdk/aws-ecs-patterns/lib/index.ts
@@ -1,23 +1,23 @@
-export * from './ecs/queue-processing-ecs-service';
+export * from './ec2/queue-processing-ecs-service';
 export * from './fargate/queue-processing-fargate-service';
 export * from './base/queue-processing-service-base';
 
-export * from './ecs/network-load-balanced-ecs-service';
+export * from './ec2/network-load-balanced-ecs-service';
 export * from './fargate/network-load-balanced-fargate-service';
 export * from './base/network-load-balanced-service-base';
 
-export * from './ecs/application-load-balanced-ecs-service';
+export * from './ec2/application-load-balanced-ecs-service';
 export * from './fargate/application-load-balanced-fargate-service';
 export * from './base/application-load-balanced-service-base';
 
-export * from './ecs/scheduled-ecs-task';
+export * from './ec2/scheduled-ecs-task';
 export * from './fargate/scheduled-fargate-task';
 export * from './base/scheduled-task-base';
 
 export * from './base/application-multiple-target-groups-service-base';
-export * from './ecs/application-multiple-target-groups-ecs-service';
+export * from './ec2/application-multiple-target-groups-ecs-service';
 export * from './fargate/application-multiple-target-groups-fargate-service';
 
 export * from './base/network-multiple-target-groups-service-base';
-export * from './ecs/network-multiple-target-groups-ecs-service';
+export * from './ec2/network-multiple-target-groups-ecs-service';
 export * from './fargate/network-multiple-target-groups-fargate-service';
